+++ CHANGES.md
@@ -1,9 +1,10 @@
 # Change Log
 
-#### next release (8.3.7)
+#### next release (8.4.0)
 
 - **Breaking change:** Replaced `node-sass` with (dart) `sass`
   - You will need to update your `TerriaMap` to use `sass` instead of `node-sass`.
+- Fix `WebMapServiceCatalogItem` `allowFeaturePicking`
 - Allow translation of TableStylingWorkflow.
 - Fix "Remove all" not removing selected/picked features
 - [The next improvement]